+++ mureq.py
@@ -5,6 +5,7 @@ in-tree by Linux systems software and other lightweight applications.
 mureq is copyright 2021 by its contributors and is released under the
 0BSD ("zero-clause BSD") license.
 """
+# fmt: off
 import contextlib
 import io
 import os.path
@@ -12,20 +13,24 @@ import socket
 import ssl
 import sys
 import urllib.parse
-from http.client import HTTPConnection, HTTPSConnection, HTTPMessage, HTTPException
+from collections.abc import Generator, MutableMapping
+from http.client import HTTPConnection, HTTPSConnection, HTTPMessage, HTTPException, HTTPResponse
+from typing import Any, cast
 
-__version__ = '0.2.0'
+__version__ = '0.3.0'
 
 __all__ = ['HTTPException', 'TooManyRedirects', 'Response',
            'yield_response', 'request', 'get', 'post', 'head', 'put', 'patch', 'delete']
 
-DEFAULT_TIMEOUT = 15.0
+DEFAULT_TIMEOUT: float = 15.0
 
 # e.g. "Python 3.8.10"
-DEFAULT_UA = "Python " + sys.version.split()[0]
+DEFAULT_UA: str = "Python " + sys.version.split()[0]
+
+Headers = MutableMapping[str, str] | HTTPMessage
 
 
-def request(method, url, *, read_limit=None, **kwargs):
+def request(method: str, url: str, *, read_limit: int | None = None,  **kwargs) -> "Response":
     """request performs an HTTP request and reads the entire response body.
 
     :param str method: HTTP method to request (e.g. 'GET', 'POST')
@@ -42,45 +47,59 @@ def request(method, url, *, read_limit=None, **kwargs):
             body = response.read(read_limit)
         except HTTPException:
             raise
-        except IOError as e:
+        except OSError as e:
             raise HTTPException(str(e)) from e
-        return Response(response.url, response.status, _prepare_incoming_headers(response.headers), body)
+        headers, raw_headers = _prepare_incoming_headers(response.headers)
+        return Response(response.url, response.status, headers, raw_headers, body)
 
 
-def get(url, **kwargs):
+def get(url: str, **kwargs) -> "Response":
     """get performs an HTTP GET request."""
     return request('GET', url=url, **kwargs)
 
 
-def post(url, body=None, **kwargs):
+def post(url: str, body: bytes | None = None, **kwargs) -> "Response":
     """post performs an HTTP POST request."""
     return request('POST', url=url, body=body, **kwargs)
 
 
-def head(url, **kwargs):
+def head(url: str, **kwargs) -> "Response":
     """head performs an HTTP HEAD request."""
     return request('HEAD', url=url, **kwargs)
 
 
-def put(url, body=None, **kwargs):
+def put(url: str, body: bytes | None = None, **kwargs) -> "Response":
     """put performs an HTTP PUT request."""
     return request('PUT', url=url, body=body, **kwargs)
 
 
-def patch(url, body=None, **kwargs):
+def patch(url: str, body: bytes | None = None, **kwargs) -> "Response":
     """patch performs an HTTP PATCH request."""
     return request('PATCH', url=url, body=body, **kwargs)
 
 
-def delete(url, **kwargs):
+def delete(url: str, **kwargs) -> "Response":
     """delete performs an HTTP DELETE request."""
     return request('DELETE', url=url, **kwargs)
 
 
 @contextlib.contextmanager
-def yield_response(method, url, *, unix_socket=None, timeout=DEFAULT_TIMEOUT, headers=None,
-                   params=None, body=None, form=None, json=None, verify=True, source_address=None,
-                   max_redirects=None, ssl_context=None):
+def yield_response(
+    method: str,
+    url: str,
+    *,
+    unix_socket: str | None = None,
+    timeout: float | None = DEFAULT_TIMEOUT,
+    headers: Headers | list[tuple[str, str]] | None = None,
+    params: dict[str, str | bytes] | list[tuple[str, str | bytes]] | None = None,
+    body: bytes | None = None,
+    form: dict[str, str | bytes] | list[tuple[str, str | bytes]] | None = None,
+    json: Any = None,
+    verify: bool = True,
+    source_address: str | tuple[str, int] | None = None,
+    max_redirects: int | None = None,
+    ssl_context: ssl.SSLContext | None = None,
+) -> Generator[HTTPResponse, None, None]:
     """yield_response is a low-level API that exposes the actual
     http.client.HTTPResponse via a contextmanager.
 
@@ -118,7 +137,7 @@ def yield_response(method, url, *, unix_socket=None, timeout=DEFAULT_TIMEOUT, he
     enc_params = _prepare_params(params)
     body = _prepare_body(body, form, json, headers)
 
-    visited_urls = []
+    visited_urls: list[str] = []
 
     while max_redirects is None or len(visited_urls) <= max_redirects:
         url, conn, path = _prepare_request(method, url, enc_params=enc_params, timeout=timeout, unix_socket=unix_socket, verify=verify, source_address=source_address, ssl_context=ssl_context)
@@ -126,12 +145,12 @@ def yield_response(method, url, *, unix_socket=None, timeout=DEFAULT_TIMEOUT, he
         visited_urls.append(url)
         try:
             try:
-                conn.request(method, path, headers=headers, body=body)
+                conn.request(method, path, headers=cast(Any, headers), body=body)
                 response = conn.getresponse()
             except HTTPException:
                 raise
-            except IOError as e:
-                # wrap any IOError that is not already an HTTPException
+            except OSError as e:
+                # wrap any OSError that is not already an HTTPException
                 # in HTTPException, exposing a uniform API for remote errors
                 raise HTTPException(str(e)) from e
             redirect_url = _check_redirect(url, response.status, response.headers)
@@ -144,6 +163,7 @@ def yield_response(method, url, *, unix_socket=None, timeout=DEFAULT_TIMEOUT, he
                 if response.status == 303:
                     # 303 See Other: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/303
                     method = 'GET'
+                    body = None
         finally:
             conn.close()
 
@@ -156,41 +176,48 @@ class Response:
     :ivar str url: the retrieved URL, indicating whether a redirection occurred
     :ivar int status_code: the HTTP status code
     :ivar http.client.HTTPMessage headers: the HTTP headers
+    :ivar raw_headers: the original unmerged HTTP headers as a list of tuples
     :ivar bytes body: the payload body of the response
     """
 
-    __slots__ = ('url', 'status_code', 'headers', 'body')
+    __slots__ = ('url', 'status_code', 'headers', 'raw_headers', 'body')
+    url: str
+    status_code: int
+    headers: Headers
+    raw_headers: list[tuple[str, str]]
+    body: bytes
 
-    def __init__(self, url, status_code, headers, body):
-        self.url, self.status_code, self.headers, self.body = url, status_code, headers, body
+    def __init__(self, url: str, status_code: int, headers: Headers, raw_headers: list[tuple[str, str]], body: bytes):
+        self.url, self.status_code, self.headers, self.raw_headers, self.body = \
+            url, status_code, headers, raw_headers, body
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return f"Response(status_code={self.status_code:d})"
 
     @property
-    def ok(self):
+    def ok(self) -> bool:
         """ok returns whether the response had a successful status code
         (anything other than a 40x or 50x)."""
         return not (400 <= self.status_code < 600)
 
     @property
-    def content(self):
+    def content(self) -> bytes:
         """content returns the response body (the `body` member). This is an
         alias for compatibility with requests.Response."""
         return self.body
 
-    def raise_for_status(self):
+    def raise_for_status(self) -> None:
         """raise_for_status checks the response's success code, raising an
         exception for error codes."""
         if not self.ok:
             raise HTTPErrorStatus(self.status_code)
 
-    def json(self):
+    def json(self) -> Any:
         """Attempts to deserialize the response body as UTF-8 encoded JSON."""
         import json as jsonlib
         return jsonlib.loads(self.body)
 
-    def _debugstr(self):
+    def _debugstr(self) -> str:
         buf = io.StringIO()
         print("HTTP", self.status_code, file=buf)
         for k, v in self.headers.items():
@@ -216,10 +243,10 @@ class HTTPErrorStatus(HTTPException):
     called explicitly.
     """
 
-    def __init__(self, status_code):
+    def __init__(self, status_code: int):
         self.status_code = status_code
 
-    def __str__(self):
+    def __str__(self) -> str:
         return f"HTTP response returned error code {self.status_code:d}"
 
 
@@ -249,7 +276,7 @@ class UnixHTTPConnection(HTTPConnection):
         self.sock = sock
 
 
-def _check_redirect(url, status, response_headers):
+def _check_redirect(url: str, status: int, response_headers: HTTPMessage) -> str | None:
     """Return the URL to redirect to, or None for no redirection."""
     if status not in (301, 302, 303, 307, 308):
         return None
@@ -276,7 +303,7 @@ def _check_redirect(url, status, response_headers):
                                     parsed_location.query, parsed_location.fragment))
 
 
-def _prepare_outgoing_headers(headers):
+def _prepare_outgoing_headers(headers: Headers | list[tuple[str, str]] | None) -> HTTPMessage:
     if headers is None:
         headers = HTTPMessage()
     elif not isinstance(headers, HTTPMessage):
@@ -294,16 +321,18 @@ def _prepare_outgoing_headers(headers):
 
 # XXX join multi-headers together so that get(), __getitem__(),
 # etc. behave intuitively, then stuff them back in an HTTPMessage.
-def _prepare_incoming_headers(headers):
-    headers_dict = {}
+def _prepare_incoming_headers(headers: HTTPMessage) -> tuple[HTTPMessage, list[tuple[str, str]]]:
+    raw_headers: list[tuple[str, str]] = []
+    headers_dict: dict[str, list[str]] = {}
     for k, v in headers.items():
         headers_dict.setdefault(k, []).append(v)
+        raw_headers.append((k, v))
     result = HTTPMessage()
     # note that iterating over headers_dict preserves the original
     # insertion order in all versions since Python 3.6:
     for k, vlist in headers_dict.items():
-        result[k] = ','.join(vlist)
-    return result
+        result[k] = ', '.join(vlist)
+    return result, raw_headers
 
 
 def _setdefault_header(headers, name, value):
@@ -311,7 +340,7 @@ def _setdefault_header(headers, name, value):
         headers[name] = value
 
 
-def _prepare_body(body, form, json, headers):
+def _prepare_body(body, form, json, headers) -> bytes | None:
     if body is not None:
         if not isinstance(body, bytes):
             raise TypeError('body must be bytes or None', type(body))
@@ -324,7 +353,7 @@ def _prepare_body(body, form, json, headers):
 
     if form is not None:
         _setdefault_header(headers, 'Content-Type', _FORM_CONTENTTYPE)
-        return urllib.parse.urlencode(form, doseq=True)
+        return urllib.parse.urlencode(form, doseq=True).encode('ascii')
 
     return None
 
@@ -352,6 +381,8 @@ def _prepare_request(method, url, *, enc_params='', timeout=DEFAULT_TIMEOUT, sou
 
     is_https = (scheme == 'https')
     host = parsed_url.hostname
+    if host is None:
+        raise ValueError("host is missing from url", url)
     port = 443 if is_https else 80
     if parsed_url.port:
         port = parsed_url.port